+++ docs/CHANGELOG.md
@@ -33,6 +33,7 @@
 - Fixed crash when immediately pressing enter after loading the UI when the `on_startup` config is set to `url` (this was due to lazy loading of the UI, and attempting to send a request before the UI was fully loaded).
 - Fixed accepting a completion via enter in the header editor also adding the header.
     - Now, the first enter press will accept the completion, and the second enter press will add the header.
+- Fixed Textual markup not being escaped key value tables.
 
 ## 2.5.4 [13th March 2025]